RESOLVED - issue SPAGOBI-722: Show the list of last N inputed value on request to the user https://spagobi.eng.it/jira/browse/SPAGOBI-722
Note: At the begin we have implemented option 2 using the combination ctrl-arrowup & ctrl-arrowdown to move in the list of cached values. Now we have modified it in order to to make the use of ctrl key optional. In order to activate or deactivate ctrl it is possible to use the setting variable moveInMementoUsingCtrlKey in file Settings.js. By default it is set to false so the user can move in the list of cached values just using arrowup and arrowdown keys.

git-svn-id: svn://svn.forge.objectweb.org/svnroot/spagobi/V_3.x/Server/trunk@17235 99afaf0d-6903-0410-885a-c66a8bbb5f81